+++ packages/tendermint-rpc/src/tendermint33/index.ts
@@ -1,6 +1,7 @@
-export { adaptor33, adaptor34 } from "./adaptors";
-export { Adaptor } from "./adaptor";
-export { Client } from "./client";
+// Note: all exports in this module are public available via
+// `import { tendermint33 } from "@cosmjs/tendermint-rpc"`
+
+export { Tendermint33Client } from "./tendermint33client";
 export {
   AbciInfoRequest,
   AbciQueryParams,