+++ packages/tendermint-rpc/src/tendermint33/tendermint33client.spec.ts
@@ -16,17 +16,16 @@ import {
   tendermintInstances,
   tendermintSearchIndexUpdated,
 } from "../testutil.spec";
-import { Adaptor } from "./adaptor";
-import { adaptorForVersion } from "./adaptors";
-import { Client } from "./client";
+import { adaptor33 } from "./adaptor";
 import { buildQuery } from "./requests";
 import * as responses from "./responses";
+import { Tendermint33Client } from "./tendermint33client";
 
-function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expected: ExpectedValues): void {
+function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues): void {
   describe("create", () => {
     it("can auto-discover Tendermint version and communicate", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory());
+      const client = await Tendermint33Client.create(rpcFactory());
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();
@@ -34,7 +33,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can connect to Tendermint with known version", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
       expect(await client.abciInfo()).toBeTruthy();
       client.disconnect();
     });
@@ -42,7 +41,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get genesis", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const genesis = await client.genesis();
     expect(genesis).toBeTruthy();
     client.disconnect();
@@ -50,7 +49,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can broadcast a transaction", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const tx = buildKvTx(randomString(), randomString());
 
     const response = await client.broadcastTxCommit({ tx: tx });
@@ -68,9 +67,9 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("gets the same tx hash from backend as calculated locally", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const tx = buildKvTx(randomString(), randomString());
-    const calculatedTxHash = adaptor.hashTx(tx);
+    const calculatedTxHash = adaptor33.hashTx(tx);
 
     const response = await client.broadcastTxCommit({ tx: tx });
     expect(response.hash).toEqual(calculatedTxHash);
@@ -80,7 +79,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can query the state", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
 
     const key = randomString();
     const value = randomString();
@@ -99,7 +98,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get a commit", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const response = await client.commit(4);
 
     expect(response).toBeTruthy();
@@ -114,7 +113,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get validators", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const response = await client.validators({});
 
     expect(response).toBeTruthy();
@@ -132,7 +131,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get all validators", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const response = await client.validatorsAll();
 
     expect(response).toBeTruthy();
@@ -150,7 +149,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can call a bunch of methods", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
 
     expect(await client.block()).toBeTruthy();
     expect(await client.genesis()).toBeTruthy();
@@ -162,7 +161,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("status", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const status = await client.status();
 
@@ -192,7 +191,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockResults", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const height = 3;
       const results = await client.blockResults(height);
@@ -208,7 +207,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockchain", () => {
     it("returns latest in descending order by default", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       // Run in parallel to increase chance there is no block between the calls
       const [status, blockchain] = await Promise.all([client.status(), client.blockchain()]);
@@ -225,7 +224,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height - 1);
@@ -239,7 +238,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("works with maxHeight in the future", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height + 20);
@@ -253,7 +252,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by minHeight and maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 2, height - 1);
@@ -267,7 +266,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("contains all the info", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 1, height - 1);
@@ -297,7 +296,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("tx", () => {
     it("can query a tx properly", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const find = randomString();
       const me = randomString();
@@ -354,7 +353,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     beforeAll(async () => {
       if (tendermintEnabled()) {
-        const client = await Client.create(rpcFactory(), adaptor);
+        const client = await Tendermint33Client.create(rpcFactory());
 
         // eslint-disable-next-line no-inner-declarations
         async function sendTx(): Promise<void> {
@@ -380,7 +379,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can paginate over txSearch results", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -399,7 +398,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can get all search results in one call", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -416,7 +415,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   });
 }
 
-function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expected: ExpectedValues): void {
+function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues): void {
   it("can subscribe to block header events", (done) => {
     pendingWithoutTendermint();
 
@@ -424,7 +423,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
 
     (async () => {
       const events: responses.NewBlockHeaderEvent[] = [];
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint33Client.create(rpcFactory());
       const stream = client.subscribeNewBlockHeader();
       expect(stream).toBeTruthy();
       const subscription = stream.subscribe({
@@ -482,7 +481,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.NewBlockEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const stream = client.subscribeNewBlock();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -538,7 +537,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     pendingWithoutTendermint();
 
     const events: responses.TxEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const stream = client.subscribeTx();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -582,7 +581,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.TxEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const query = buildQuery({ tags: [{ key: "app.creator", value: expected.appCreator }] });
     const stream = client.subscribeTx(query);
     expect(stream).toBeTruthy();
@@ -620,7 +619,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can unsubscribe and re-subscribe to the same stream", async () => {
     pendingWithoutTendermint();
 
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const stream = client.subscribeNewBlockHeader();
 
     const event1 = await firstEvent(stream);
@@ -653,7 +652,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can subscribe twice", async () => {
     pendingWithoutTendermint();
 
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint33Client.create(rpcFactory());
     const stream1 = client.subscribeNewBlockHeader();
     const stream2 = client.subscribeNewBlockHeader();
 
@@ -665,48 +664,46 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   });
 }
 
-for (const { url, version, expected } of tendermintInstances) {
-  describe(`Client ${version}`, () => {
-    it("can connect to a given url", async () => {
-      pendingWithoutTendermint();
+describe(`Tendermint33Client`, () => {
+  const { url, expected } = tendermintInstances[0];
 
-      // default connection
-      {
-        const client = await Client.connect(url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
+  it("can connect to a given url", async () => {
+    pendingWithoutTendermint();
 
-      // http connection
-      {
-        const client = await Client.connect("http://" + url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
+    // default connection
+    {
+      const client = await Tendermint33Client.connect(url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
 
-      // ws connection
-      {
-        const client = await Client.connect("ws://" + url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
-    });
+    // http connection
+    {
+      const client = await Tendermint33Client.connect("http://" + url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
 
-    describe("With HttpClient", () => {
-      const adaptor = adaptorForVersion(version);
-      defaultTestSuite(() => new HttpClient(url), adaptor, expected);
-    });
-
-    describe("With WebsocketClient", () => {
-      // don't print out WebSocket errors if marked pending
-      const onError = process.env.TENDERMINT_ENABLED ? console.error : () => 0;
-      const factory = (): WebsocketClient => new WebsocketClient(url, onError);
-      const adaptor = adaptorForVersion(version);
-      defaultTestSuite(factory, adaptor, expected);
-      websocketTestSuite(factory, adaptor, expected);
-    });
+    // ws connection
+    {
+      const client = await Tendermint33Client.connect("ws://" + url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
   });
-}
+
+  describe("With HttpClient", () => {
+    defaultTestSuite(() => new HttpClient(url), expected);
+  });
+
+  describe("With WebsocketClient", () => {
+    // don't print out WebSocket errors if marked pending
+    const onError = process.env.TENDERMINT_ENABLED ? console.error : () => 0;
+    const factory = (): WebsocketClient => new WebsocketClient(url, onError);
+    defaultTestSuite(factory, expected);
+    websocketTestSuite(factory, expected);
+  });
+});