+++ packages/tendermint-rpc/src/tendermint33/tendermint33client.ts
@@ -9,38 +9,32 @@ import {
   SubscriptionEvent,
   WebsocketClient,
 } from "../rpcclients";
-import { Adaptor, Decoder, Encoder, Params, Responses } from "./adaptor";
-import { adaptorForVersion } from "./adaptors";
+import { adaptor33, Decoder, Encoder, Params, Responses } from "./adaptor";
 import * as requests from "./requests";
 import * as responses from "./responses";
 
-/** @deprecated Use Tendermint34Client */
-export class Client {
+export class Tendermint33Client {
   /**
    * Creates a new Tendermint client for the given endpoint.
    *
    * Uses HTTP when the URL schema is http or https. Uses WebSockets otherwise.
-   *
-   * If the adaptor is not set an auto-detection is attempted.
    */
-  public static async connect(url: string, adaptor?: Adaptor): Promise<Client> {
+  public static async connect(url: string): Promise<Tendermint33Client> {
     const useHttp = url.startsWith("http://") || url.startsWith("https://");
     const rpcClient = useHttp ? new HttpClient(url) : new WebsocketClient(url);
-    return Client.create(rpcClient, adaptor);
+    return Tendermint33Client.create(rpcClient);
   }
 
   /**
    * Creates a new Tendermint client given an RPC client.
-   *
-   * If the adaptor is not set an auto-detection is attempted.
    */
-  public static async create(rpcClient: RpcClient, adaptor?: Adaptor): Promise<Client> {
+  public static async create(rpcClient: RpcClient): Promise<Tendermint33Client> {
     // For some very strange reason I don't understand, tests start to fail on some systems
     // (our CI) when skipping the status call before doing other queries. Sleeping a little
     // while did not help. Thus we query the version as a way to say "hi" to the backend,
     // even in cases where we don't use the result.
-    const version = await this.detectVersion(rpcClient);
-    return new Client(rpcClient, adaptor || adaptorForVersion(version));
+    const _version = await this.detectVersion(rpcClient);
+    return new Tendermint33Client(rpcClient);
   }
 
   private static async detectVersion(client: RpcClient): Promise<string> {
@@ -64,12 +58,12 @@ export class Client {
   private readonly r: Responses;
 
   /**
-   * Use `Client.connect` or `Client.create` to create an instance.
+   * Use `Tendermint33Client.connect` or `Tendermint33Client.create` to create an instance.
    */
-  private constructor(client: RpcClient, adaptor: Adaptor) {
+  private constructor(client: RpcClient) {
     this.client = client;
-    this.p = adaptor.params;
-    this.r = adaptor.responses;
+    this.p = adaptor33.params;
+    this.r = adaptor33.responses;
   }
 
   public disconnect(): void {